changed interface to repos. cleaned up models

+++ src/main/java/com/github/mwolf1963/improvementtracker/controllers/DepartmentController.java
@@ -36,7 +36,7 @@ public class DepartmentController {
     }
     
     @GetMapping("/{id}")
-    public Department get(@PathVariable("id") long id){
+    public Department get(@PathVariable("id") int id){
         return departmentRepository.getOne(id);
     }
