+++ src/main/java/com/github/mwolf1963/improvementtracker/repositories/MaterialRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.github.mwolf1963.improvementtracker.models.Material;
 
-public interface MaterialRepository extends JpaRepository<Material, Long> {
+public interface MaterialRepository extends JpaRepository<Material, Integer> {
 
 }
