+++ src/main/java/com/github/mwolf1963/improvementtracker/repositories/PartRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.github.mwolf1963.improvementtracker.models.Part;
 
-public interface PartRepository extends JpaRepository<Part, Long> {
+public interface PartRepository extends JpaRepository<Part, Integer> {
 
 }
